Refactored alerts logic into own file
