+++ src/store/AlertStore.ts
@@ -1,0 +1,19 @@
+import { proxyMap } from "valtio/utils";
+
+
+export interface AlertProps {
+  type: "success" | "error" | "info" | "loading";
+  title?: string;
+  message: string;
+  key: string;
+  closable?: boolean;
+  duration?: number;
+}
+export const alertsStore = proxyMap<string, AlertProps>(); // Map to ensure insertion order
+
+export const pushAlert = (alert: AlertProps) => {
+  alertsStore.set(alert.key, alert);
+};
+export const removeAlert = (key: string) => {
+  alertsStore.delete(key);
+};
