+++ src/store/index.ts
@@ -1,6 +1,5 @@
 import { Chains } from '@reactive-dot/core';
 import { proxy } from 'valtio'
-import { proxyMap } from "valtio/utils"
 import { config } from '~/api/config';
 
 export * from './userSore'
@@ -33,18 +32,3 @@ export const appStore = proxy<AppStore>({
   isDevTest: import.meta.env.MODE === 'development',
 })
 
-export interface AlertProps {
-  type: "success" | "error" | "info" | "loading";
-  title?: string;
-  message: string;
-  key: string;
-  closable?: boolean;
-  duration?: number;
-}
-export const alertsStore = proxyMap<string, AlertProps>();  // Map to ensure insertion order
-export const pushAlert = (alert: AlertProps) => {
-  alertsStore.set(alert.key, alert);
-}
-export const removeAlert = (key: string) => {
-  alertsStore.delete(key);
-}
